simplify MetaDatasource example
