--- great_expectations/zep/advanced_bootstrap_poc/gx_sql.py
@@ -1,6 +1,0 @@
-from great_expectations.zep.core import SQLDatasource
-
-
-def add_sql(self, name: str) -> SQLDatasource:
-    print(f"Added SQL - {name}")
-    return SQLDatasource()
