--- great_expectations/zep/metaclass_poc/config.yaml
@@ -1,13 +1,0 @@
-datasources:
-  taxi_datasource:
-    version: 2.0
-    engine: pandas
-    assets:
-      yellow_trip_data:
-        type: file
-        base_directory: ../data
-        separators: \t
-        has_headers: false
-      yellow_trip_data_but_sql:
-        type: sql
-        query: "SELECT * FROM TAXI"
