--- great_expectations/zep/sources_module_ns_poc/gx2_sql.py
@@ -1,6 +1,0 @@
-from great_expectations.zep.core import SQLDatasource
-
-
-def add_sql(name: str) -> SQLDatasource:
-    print(f"Added SQL - {name}")
-    return SQLDatasource()
